pyproject.toml
[project]
name = "kubernetes-training"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "apache-airflow==2.10.5",
    "fastapi>=0.120.2",
    "pandas==1.5.3",
    "psycopg2-binary==2.9.9",
    "uvicorn>=0.38.0",
]
